requirements.txt
# Core
pydantic>=2.0
pydantic-settings
python-dotenv

# API Layer (FastAPI)
fastapi
uvicorn[standard]
python-multipart

# UI Layer (Streamlit)
streamlit
watchdog

# AI/ML Engines (Optimized)
torch --index-url https://download.pytorch.org/whl/cu118
transformers
accelerate
bitsandbytes
scipy
soundfile
librosa

# Google Gemini
google-generativeai

# ASR Engine (WhisperX - faster than openai-whisper)
whisperx
noisereduce
faster-whisper

# Utilities
requests
pyyaml
tqdm

# Vector RAG (FR-02)
chromadb
sentence-transformers

# Async Queue & Workers (Production)
arq
redis
async-timeout

# Database & ORM
sqlmodel
alembic
asyncpg
aiosqlite

# Ops & Performance
psutil
orjson
slowapi

# Networking
httpx
